feat(chatgpt): add chatgpt.go file
Add a new file  to the  directory. This file contains the implementation of a function  which generates a commit message based on user input.

The function uses the Go-ChatGPT SDK and the OpenAI GPT-3.5-Turbo model to generate the commit message. It takes a string  as input and returns a  and an error.

The generated commit message will follow the karma style format: . The commit message should be converted to lowercase format.

To generate the commit message, the function initializes the SDK configuration with the OpenAI API key obtained from the environment variable .

A system message is first sent to provide instructions to the model. The user message, which contains the output of the diff --git a/internal/chatgpt/chatgpt.go b/internal/chatgpt/chatgpt.go
new file mode 100644
index 0000000..d4de517
--- /dev/null
+++ b/internal/chatgpt/chatgpt.go
@@ -0,0 +1,35 @@
+package chatgpt
+
+import (
+	"os"
+
+	sdk "github.com/ak9024/go-chatgpt-sdk"
+)
+
+// generatedCommitMessageByChatGPT(c string) to generate commit message by input in param `c`
+func GeneratedCommitMessageByChatGPT(c string) (*sdk.ModelChatResponse, error) {
+	_sdk := sdk.NewConfig(sdk.Config{
+		OpenAIKey: os.Getenv("OPENAI_API_KEY"),
+	})
+
+	content := "You are act as the author of a commit message in git, your mission is to create clean and comprehensive commit messages, and also desribe the why and what please follow the karma style like `<type>(type): <subject> <body> <footer>`, i'll send you output of `git diff --staged` command, and you convert it into commit message. please don't add any description to the commit, only commit message and convert all message into lowwercase format"
+
+	resp, err := _sdk.ChatCompletions(sdk.ModelChat{
+		Model: "gpt-3.5-turbo",
+		Messages: []sdk.Message{
+			{
+				Role:    "system",
+				Content: content,
+			},
+			{
+				Role:    "user",
+				Content: c,
+			},
+		},
+	})
+	if err != nil {
+		return nil, err
+	}
+
+	return resp, nil
+} command, is then sent to the model.

The response from the model, which includes the generated commit message, is returned.

Note: The commit message should not include any description, only the commit message itself should be provided.