chore(cmd): add constant.go
Add a new file constant.go to the cmd package. This file contains constants used in the command line interface (CLI) for generating commits automatically.

The constants include:
- cliVersion: the current version of the CLI
- shortText: a short description of the CLI
- longText: a longer description of the CLI and its usage

This file will be used to store constant values that are used across the CLI.

chore(cmd): update root.go

Update the root.go file in the cmd package.

- Set the CLI name to go-commit.
- Set the CLI version to the value stored in the cliVersion constant.
- Set the short description of the CLI to the value stored in the shortText constant.
- Set the long description of the CLI to the value stored in the Long constant.

These updates ensure that the CLI is properly named and versioned, and that