feat(commit): add success message after commit
Add a success message after committing a change. This message will be printed if the commit is successful. The success message informs the user to check the go log for further details.

The commit logic has been updated in the  function in the  file. After executing the commit command, a success message will be printed if the  variable is not empty.